Add image-mode cycling (m) and clear out checkup findings
Wires up the long-planned `m` keybinding to cycle the image rendering
mode (full/block/geo/ascii) in both static and animated image views,
with the active mode shown in the status line. Removes the now-dead
`Background::label()`, the redundant `as i32` casts in gen_glyphs, the
sub-MSRV `is_multiple_of` call, and a needless_range_loop in clustering
tests. Drops the checkup doc whose findings these resolve.

+++ examples/gen_glyphs.rs
@@ -190,8 +190,8 @@ fn compose_glyph_on_canvas(
     // - glyph top edge = ascent - (ymin + height)
     // - glyph left edge = xmin
 
-    let glyph_top = ascent as i32 - metrics.ymin as i32 - metrics.height as i32;
-    let glyph_left = metrics.xmin as i32;
+    let glyph_top = ascent as i32 - metrics.ymin - metrics.height as i32;
+    let glyph_left = metrics.xmin;
 
     for gy in 0..metrics.height {
         for gx in 0..metrics.width {
